+++ bin/corpse_tools.dart
@@ -1,20 +1,22 @@
-import 'dart:io';
-
 import 'package:args/command_runner.dart';
 import 'package:corpse_tools/commands.dart';
 
-void main(List<String> args) {
-  var desc = 'CorpseTools v0.1.1 made by Kyouin '
+void main(List<String> args) async {
+  var desc = 'CorpseTools v0.1.3 made by Kyouin '
       '(https://github.com/KyouinDev/corpse_tools)\n'
-      'Supported game(s):\n'
-      ' - Corpse Party: Book of Shadows (PC)';
+      'Supported games:\n'
+      ' - Corpse Party: Book of Shadows (2018, PC)'
+      ' - Corpse Party: Sweet Sachiko\'s Hysteric Birthday Bash (2019, PC)';
   var runner = CommandRunner('corpse_tools.exe', desc)
     ..addCommand(ExtractTextCommand())
     ..addCommand(ReplaceTextCommand());
-  runner.run(args).catchError((error) {
-    if (error is! UsageException) throw error;
 
-    print(error);
-    exit(-1);
-  });
+  try {
+    await runner.run(args);
+  } on UsageException catch(e, s) {
+    print('$e\n$s');
+  } on Exception catch(e, s) {
+    print('Unknown $e\n$s');
+    print('Please report this to the developer.');
+  }
 }