+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3
+
+- Confirmed support for "Corpse Party: Sweet Sachiko's Hysteric Birthday Bash".
+- Optimized main class.
+
 ## 0.1.2
 
 - Application no longer depends on definitions file.